+++ syntropy/knn/__init__.py
@@ -2,6 +2,7 @@ from .shannon import (
     differential_entropy,
     mutual_information,
     conditional_mutual_information,
+    kullback_leibler_divergence
 )
 
 from .multivariate_mi import (
@@ -20,6 +21,7 @@ __all__ = [
     "differential_entropy",
     "mutual_information",
     "conditional_mutual_information",
+    "kullback_leibler_divergence",
     "total_correlation",
     "dual_total_correlation",
     "s_information",